Clear your CatastRoEus cache dir

Use this function with caution. This function would clear your cached data and configuration, specifically:

Deletes the CatastRoEus config directory (rappdirs::user_config_dir("CatastRoEus", "R")).

Deletes the cache_dir directory.

Deletes the values on stored on Sys.getenv("CATASTROEUS_CACHE_DIR").

catreus_clear_cache(config = FALSE, cached_data = TRUE, verbose = FALSE)

Arguments

config

if TRUE, will delete the configuration folder of CatastRoEus.

cached_data

If this is set to TRUE, it will delete your cache_dir and all its content.

verbose

Logical, displays information. Useful for debugging, default is FALSE.

Value

Invisible. This function is called for its side effects.

Details

This is an overkill function that is intended to reset your status as it you would never have installed and/or used CatastRoEus.

See also

Examples

# Don't run this! It would modify your current state
# \dontrun{
catreus_clear_cache(verbose = TRUE)
#> CatastRoEus cached data deleted: C:\Users\RUNNER~1\AppData\Local\Temp\RtmpGmLmbg/CatastRoEus
# }

Sys.getenv("CATASTROEUS_CACHE_DIR")
#> [1] ""